ref: PlayerDisplayで不要な情報をpropとして受け取らないようにする
